refatora codigo do reprocessamento e salva um documento de status de reprocessamento no mongo
